requirements.txt
openai
pillow
pytesseract
pyautogui
keyboard
python-dotenv
